+++ game/src/images.lua
@@ -5,6 +5,7 @@ function Images:load()
     self.player = love.graphics.newImage("assets/mikroga_white.png")
     self.enemybullet = love.graphics.newImage("assets/enemy_bullet_white.png")
     self.enemy1 = love.graphics.newImage("assets/enemy1_white.png")
+    self.enemy2 = love.graphics.newImage("assets/enemy2_white.png")
 end
 
 return Images